fix(package): Fix deployed package (#5)
The previous version of `@benmvp/cli` depending on `core-js` in weird ways that caused problems with `url-lib@3.0.0`.

Lots of fixes were made in `@benmvp/cli` to fix this problem. v4 also introduced the `integrate` command so that integration tests can be run before releasing. A set of integration tests has been added for each utility mirroring what is in the examples.

Lastly, switched from Yarn over to npm.