fix(types): Export TypeScript types
When `formatUrl` (and probably other utilities) are wrapped in functions, those functions will likely take arguments that will be passed through to the utilities. This means that those arguments will need to be typed correctly. Typing as `Object` will not work, it needs to be `UrlParams` specifically.

So this PR just exports `UrlParams`, `UrlValue` & `NullableUrlParams` so that they can be used as necessary. I added an integration test for this as well. The test failed with `Object` and passes with the newly exported `UrlParams`.